test(codegen): drop legacy ID-facade template tests
TestLegacyGoSDKInterfaceSurface and TestLegacyTypeScriptIDFacade were
written against the template-driven legacy compat (per-type FooID
aliases, manual loadFooFromID wrappers), which has been moved to the
engine schema via view-filtered scalar installation. The remaining
slivers of legacy codegen logic (Go interface-as-struct, TS
@expectedType→FooID rewrites) are exercised end-to-end by
TestLegacy/*FromIDCompat. Rename the Go file to reflect that only the
modern surface test remains.
